Added solution

+++ app/controllers/toys_controller.rb
@@ -16,10 +16,24 @@ class ToysController < ApplicationController
     toy.update(toy_params)
   end
 
-  def destroy
+  def increment_likes
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.update(likes: toy.likes + 1)
+      render json: toy
+    else
+      render json: { error: "Toy not found" }, status: :not_found
+    end
+  end
+
+  def destroy
+      toy = Toy.find_by(id: params[:id])
+    if toy
+      toy.destroy
+      head :no_content
+    else
+      render josn: { error: "Bird not found" }, status: :not_found
+    end
   end
 
   private
